refactor: give image viewing its own home
